refactor(api): rename certification-result-service to scoring-certification-service
