+++ api/lib/domain/services/scoring/scoring-certification-service.js
@@ -1,9 +1,125 @@
-const certificationResultService = require('./certification-result-service');
+const _ = require('lodash');
+const CertificationContract = require('../../models/CertificationContract');
+const scoringService = require('./scoring-service');
+const placementProfileService = require('../placement-profile-service');
+const { CertifiedLevel } = require('../../models/CertifiedLevel');
+const { CertifiedScore } = require('../../models/CertifiedScore');
+const { ReproducibilityRate } = require('../../models/ReproducibilityRate');
+const CompetenceMark = require('../../models/CompetenceMark');
+const CertificationAssessmentScore = require('../../models/CertificationAssessmentScore');
+const AnswerCollectionForScoring = require('../../models/AnswerCollectionForScoring');
 
-async function calculateCertificationAssessmentScore({ certificationAssessment, continueOnError }) {
-  return await certificationResultService.computeResult({ certificationAssessment, continueOnError });
+function _selectAnswersMatchingCertificationChallenges(answers, certificationChallenges) {
+  return answers.filter(
+    ({ challengeId }) => _.some(certificationChallenges, { challengeId }),
+  );
+}
+
+function _selectChallengesMatchingCompetences(certificationChallenges, testedCompetences) {
+  return certificationChallenges.filter(
+    ({ competenceId }) => _.some(testedCompetences, { id: competenceId }),
+  );
+}
+
+function _getSumScoreFromCertifiedCompetences(listCompetences) {
+  return _(listCompetences).map('obtainedScore').sum();
+}
+
+function _getCompetenceMarksWithCertifiedLevelAndScore(answers, listCompetences, reproducibilityRate, certificationChallenges, continueOnError, answerCollection) {
+  return listCompetences.map((competence) => {
+    const challengesForCompetence = _.filter(certificationChallenges, { competenceId: competence.id });
+    const answersForCompetence = _selectAnswersMatchingCertificationChallenges(answers, challengesForCompetence);
+
+    if (!continueOnError) {
+      CertificationContract.assertThatCompetenceHasAtLeastOneChallenge(challengesForCompetence, competence.index);
+      CertificationContract.assertThatCompetenceHasAtLeastOneAnswer(answersForCompetence, competence.index);
+      CertificationContract.assertThatEveryAnswerHasMatchingChallenge(answersForCompetence, challengesForCompetence);
+      CertificationContract.assertThatNoChallengeHasMoreThanOneAnswer(answersForCompetence, challengesForCompetence);
+    }
+
+    const certifiedLevel = CertifiedLevel.from({
+      numberOfChallenges: answerCollection.numberOfChallengesForCompetence(competence.id),
+      numberOfCorrectAnswers: answerCollection.numberOfCorrectAnswersForCompetence(competence.id),
+      numberOfNeutralizedAnswers: answerCollection.numberOfNeutralizedChallengesForCompetence(competence.id),
+      estimatedLevel: competence.estimatedLevel,
+      reproducibilityRate,
+    });
+    const certifiedScore = CertifiedScore.from({ certifiedLevel, estimatedScore: competence.pixScore });
+    return new CompetenceMark({
+      level: scoringService.getBlockedLevel(certifiedLevel.value),
+      score: scoringService.getBlockedPixScore(certifiedScore.value),
+      area_code: competence.area.code,
+      competence_code: competence.index,
+      competenceId: competence.id,
+    });
+  });
+}
+
+function _getCompetenceMarksWithFailedLevel(listCompetences) {
+  return listCompetences.map((competence) => {
+    return new CompetenceMark({
+      level: scoringService.getBlockedLevel(CertifiedLevel.invalidate().value),
+      score: scoringService.getBlockedPixScore(0),
+      area_code: competence.area.code,
+      competence_code: competence.index,
+      competenceId: competence.id,
+    });
+  });
+}
+
+function _getResult(answers, certificationChallenges, testedCompetences, continueOnError) {
+
+  if (!continueOnError) {
+    CertificationContract.assertThatWeHaveEnoughAnswers(answers, certificationChallenges);
+  }
+
+  const answerCollection = AnswerCollectionForScoring.from({ answers, challenges: certificationChallenges });
+
+  const reproducibilityRate = ReproducibilityRate.from({
+    numberOfNonNeutralizedChallenges: answerCollection.numberOfNonNeutralizedChallenges(),
+    numberOfCorrectAnswers: answerCollection.numberOfCorrectAnswers(),
+  });
+
+  if (!reproducibilityRate.isEnoughToBeCertified()) {
+    return new CertificationAssessmentScore({
+      competenceMarks: _getCompetenceMarksWithFailedLevel(testedCompetences),
+      percentageCorrectAnswers: reproducibilityRate.value,
+    });
+  }
+
+  const competenceMarks = _getCompetenceMarksWithCertifiedLevelAndScore(answers, testedCompetences, reproducibilityRate.value, certificationChallenges, continueOnError, answerCollection);
+  const scoreAfterRating = _getSumScoreFromCertifiedCompetences(competenceMarks);
+
+  if (!continueOnError) {
+    CertificationContract.assertThatScoreIsCoherentWithReproducibilityRate(scoreAfterRating, reproducibilityRate.value);
+  }
+
+  return new CertificationAssessmentScore({ competenceMarks, percentageCorrectAnswers: reproducibilityRate.value });
+}
+
+async function _getTestedCompetences({ userId, limitDate, isV2Certification }) {
+  const placementProfile = await placementProfileService.getPlacementProfile({ userId, limitDate, isV2Certification });
+  return _(placementProfile.userCompetences)
+    .filter((uc) => uc.isCertifiable())
+    .map((uc) => _.pick(uc, ['id', 'area', 'index', 'name', 'estimatedLevel', 'pixScore']))
+    .value();
 }
 
 module.exports = {
-  calculateCertificationAssessmentScore,
+  async calculateCertificationAssessmentScore({ certificationAssessment, continueOnError }) {
+    // userService.getPlacementProfile() + filter level > 0 => avec allCompetence (bug)
+    const testedCompetences = await _getTestedCompetences({
+      userId: certificationAssessment.userId,
+      limitDate: certificationAssessment.createdAt,
+      isV2Certification: certificationAssessment.isV2Certification,
+    });
+
+    // map sur challenges filtre sur competence Id - S'assurer qu'on ne travaille que sur les compétences certifiables
+    const matchingCertificationChallenges = _selectChallengesMatchingCompetences(certificationAssessment.certificationChallenges, testedCompetences);
+
+    // map sur challenges filtre sur challenge Id
+    const matchingAnswers = _selectAnswersMatchingCertificationChallenges(certificationAssessment.certificationAnswersByDate, matchingCertificationChallenges);
+
+    return _getResult(matchingAnswers, matchingCertificationChallenges, testedCompetences, continueOnError);
+  },
 };
