+++ api/lib/domain/usecases/index.js
@@ -49,7 +49,6 @@ const dependencies = {
   certificationRepository: require('../../infrastructure/repositories/certification-repository'),
   certificationCpfService: require('../../domain/services/certification-cpf-service'),
   certificationService: require('../../domain/services/certification-service'),
-  certificationResultService: require('../services/scoring/certification-result-service'),
   challengeRepository: require('../../infrastructure/repositories/challenge-repository'),
   challengeForPixAutoAnswerRepository: require('../../infrastructure/repositories/challenge-for-pix-auto-answer-repository'),
   cleaCertificationResultRepository: require('../../infrastructure/repositories/clea-certification-result-repository'),
