--- api/tests/unit/domain/services/scoring/certification-result-service_test.js
@@ -1,859 +1,0 @@
-const _ = require('lodash');
-const { expect, sinon, domainBuilder, catchErr } = require('../../../../test-helper');
-const certificationResultService = require('../../../../../lib/domain/services/scoring/certification-result-service');
-const { states } = require('../../../../../lib/domain/models/CertificationAssessment');
-const placementProfileService = require('../../../../../lib/domain/services/placement-profile-service');
-const UserCompetence = require('../../../../../lib/domain/models/UserCompetence');
-const { CertificationComputeError } = require('../../../../../lib/domain/errors');
-
-function _buildUserCompetence(competence, pixScore, estimatedLevel) {
-  return new UserCompetence({ ...competence, estimatedLevel, pixScore });
-}
-
-const pixForCompetence1 = 10;
-const pixForCompetence2 = 20;
-const pixForCompetence3 = 30;
-const pixForCompetence4 = 40;
-const UNCERTIFIED_LEVEL = -1;
-
-const correctAnswersForAllChallenges = () => _.map([
-  { challengeId: 'challenge_A_for_competence_1', result: 'ok' },
-  { challengeId: 'challenge_B_for_competence_1', result: 'ok' },
-  { challengeId: 'challenge_C_for_competence_1', result: 'ok' },
-  { challengeId: 'challenge_D_for_competence_2', result: 'ok' },
-  { challengeId: 'challenge_E_for_competence_2', result: 'ok' },
-  { challengeId: 'challenge_F_for_competence_2', result: 'ok' },
-  { challengeId: 'challenge_G_for_competence_3', result: 'ok' },
-  { challengeId: 'challenge_H_for_competence_3', result: 'ok' },
-  { challengeId: 'challenge_I_for_competence_3', result: 'ok' },
-  { challengeId: 'challenge_J_for_competence_4', result: 'ok' },
-  { challengeId: 'challenge_K_for_competence_4', result: 'ok' },
-  { challengeId: 'challenge_L_for_competence_4', result: 'ok' },
-], domainBuilder.buildAnswer);
-
-const wrongAnswersForAllChallenges = () => _.map([
-  { challengeId: 'challenge_A_for_competence_1', result: 'ko' },
-  { challengeId: 'challenge_B_for_competence_1', result: 'ko' },
-  { challengeId: 'challenge_C_for_competence_1', result: 'ko' },
-  { challengeId: 'challenge_D_for_competence_2', result: 'ko' },
-  { challengeId: 'challenge_E_for_competence_2', result: 'ko' },
-  { challengeId: 'challenge_F_for_competence_2', result: 'ko' },
-  { challengeId: 'challenge_G_for_competence_3', result: 'ko' },
-  { challengeId: 'challenge_H_for_competence_3', result: 'ko' },
-  { challengeId: 'challenge_I_for_competence_3', result: 'ko' },
-  { challengeId: 'challenge_J_for_competence_4', result: 'ko' },
-  { challengeId: 'challenge_K_for_competence_4', result: 'ko' },
-  { challengeId: 'challenge_L_for_competence_4', result: 'ko' },
-], domainBuilder.buildAnswer);
-
-const answersToHaveOnlyTheLastCompetenceFailed = () => _.map([
-  { challengeId: 'challenge_A_for_competence_1', result: 'ok' },
-  { challengeId: 'challenge_B_for_competence_1', result: 'ok' },
-  { challengeId: 'challenge_C_for_competence_1', result: 'ok' },
-  { challengeId: 'challenge_D_for_competence_2', result: 'ok' },
-  { challengeId: 'challenge_E_for_competence_2', result: 'ok' },
-  { challengeId: 'challenge_F_for_competence_2', result: 'ok' },
-  { challengeId: 'challenge_G_for_competence_3', result: 'ok' },
-  { challengeId: 'challenge_H_for_competence_3', result: 'ok' },
-  { challengeId: 'challenge_I_for_competence_3', result: 'ok' },
-  { challengeId: 'challenge_J_for_competence_4', result: 'ko' },
-  { challengeId: 'challenge_K_for_competence_4', result: 'ko' },
-  { challengeId: 'challenge_L_for_competence_4', result: 'ko' },
-], domainBuilder.buildAnswer);
-
-const answersWithReproducibilityRateLessThan80 = () => _.map([
-  { challengeId: 'challenge_A_for_competence_1', result: 'ok' },
-  { challengeId: 'challenge_B_for_competence_1', result: 'ko' },
-  { challengeId: 'challenge_C_for_competence_1', result: 'ok' },
-  { challengeId: 'challenge_D_for_competence_2', result: 'ok' },
-  { challengeId: 'challenge_E_for_competence_2', result: 'ok' },
-  { challengeId: 'challenge_F_for_competence_2', result: 'ok' },
-  { challengeId: 'challenge_G_for_competence_3', result: 'ok' },
-  { challengeId: 'challenge_H_for_competence_3', result: 'ko' },
-  { challengeId: 'challenge_I_for_competence_3', result: 'ko' },
-  { challengeId: 'challenge_J_for_competence_4', result: 'ok' },
-  { challengeId: 'challenge_K_for_competence_4', result: 'ko' },
-  { challengeId: 'challenge_L_for_competence_4', result: 'ok' },
-], domainBuilder.buildAnswer);
-
-const challenges = _.map([
-  { challengeId: 'challenge_A_for_competence_1', competenceId: 'competence_1', associatedSkillName: '@skillChallengeA_1', type: 'QCM' },
-  { challengeId: 'challenge_C_for_competence_1', competenceId: 'competence_1', associatedSkillName: '@skillChallengeC_1', type: 'QCM' },
-  { challengeId: 'challenge_B_for_competence_1', competenceId: 'competence_1', associatedSkillName: '@skillChallengeB_1', type: 'QCM' },
-  { challengeId: 'challenge_D_for_competence_2', competenceId: 'competence_2', associatedSkillName: '@skillChallengeD_2', type: 'QCM' },
-  { challengeId: 'challenge_E_for_competence_2', competenceId: 'competence_2', associatedSkillName: '@skillChallengeE_2', type: 'QCM' },
-  { challengeId: 'challenge_F_for_competence_2', competenceId: 'competence_2', associatedSkillName: '@skillChallengeF_2', type: 'QCM' },
-  { challengeId: 'challenge_G_for_competence_3', competenceId: 'competence_3', associatedSkillName: '@skillChallengeG_3', type: 'QCM' },
-  { challengeId: 'challenge_H_for_competence_3', competenceId: 'competence_3', associatedSkillName: '@skillChallengeH_3', type: 'QCM' },
-  { challengeId: 'challenge_I_for_competence_3', competenceId: 'competence_3', associatedSkillName: '@skillChallengeI_3', type: 'QCM' },
-  { challengeId: 'challenge_J_for_competence_4', competenceId: 'competence_4', associatedSkillName: '@skillChallengeJ_4', type: 'QCM' },
-  { challengeId: 'challenge_K_for_competence_4', competenceId: 'competence_4', associatedSkillName: '@skillChallengeK_4', type: 'QCM' },
-  { challengeId: 'challenge_L_for_competence_4', competenceId: 'competence_4', associatedSkillName: '@skillChallengeL_4', type: 'QCM' },
-], domainBuilder.buildCertificationChallengeWithType);
-
-const competence_1 = domainBuilder.buildCompetence({ id: 'competence_1', index: '1.1', area: { code: '1' }, name: 'Mener une recherche' });
-const competence_2 = domainBuilder.buildCompetence({ id: 'competence_2', index: '2.2', area: { code: '2' }, name: 'Partager' });
-const competence_3 = domainBuilder.buildCompetence({ id: 'competence_3', index: '3.3', area: { code: '3' }, name: 'Adapter' });
-const competence_4 = domainBuilder.buildCompetence({ id: 'competence_4', index: '4.4', area: { code: '4' }, name: 'Résoudre' });
-const competence_5 = domainBuilder.buildCompetence({ id: 'competence_5', index: '5.5', area: { code: '5' }, name: 'Chercher' });
-const competence_6 = domainBuilder.buildCompetence({ id: 'competence_6', index: '6.6', area: { code: '6' }, name: 'Trouver' });
-
-const userCompetences = [
-  _buildUserCompetence(competence_1, pixForCompetence1, 1),
-  _buildUserCompetence(competence_2, pixForCompetence2, 2),
-  _buildUserCompetence(competence_3, pixForCompetence3, 3),
-  _buildUserCompetence(competence_4, pixForCompetence4, 4),
-];
-
-describe('Unit | Service | Certification Result Service', function() {
-
-  context('#computeResult', () => {
-    let certificationAssessment;
-    const certificationAssessmentData = {
-      id: 1,
-      userId: 11,
-      certificationCourseId: 111,
-      createdAt: '2020-02-01T00:00:00Z',
-      completedAt: '2020-02-01T00:00:00Z',
-      state: states.COMPLETED,
-      isV2Certification: true,
-    };
-
-    const competenceWithMarks_1_1 = domainBuilder.buildCompetenceMark({
-      level: UNCERTIFIED_LEVEL,
-      score: 0,
-      area_code: '1',
-      competence_code: '1.1',
-      competenceId: 'competence_1',
-    });
-
-    const competenceWithMarks_2_2 = domainBuilder.buildCompetenceMark({
-      level: UNCERTIFIED_LEVEL,
-      score: 0,
-      area_code: '2',
-      competence_code: '2.2',
-      competenceId: 'competence_2',
-    });
-
-    const competenceWithMarks_3_3 = domainBuilder.buildCompetenceMark({
-      level: UNCERTIFIED_LEVEL,
-      score: 0,
-      area_code: '3',
-      competence_code: '3.3',
-      competenceId: 'competence_3',
-    });
-
-    const competenceWithMarks_4_4 = domainBuilder.buildCompetenceMark({
-      level: UNCERTIFIED_LEVEL,
-      score: 0,
-      area_code: '4',
-      competence_code: '4.4',
-      competenceId: 'competence_4',
-    });
-
-    const expectedCertifiedCompetences = [
-      competenceWithMarks_1_1,
-      competenceWithMarks_2_2,
-      competenceWithMarks_3_3,
-      competenceWithMarks_4_4,
-    ];
-
-    context('When at least one challenge have more than one answer ', () => {
-
-      it('should throw', async () => {
-        // given
-        const certificationAnswersByDate = _.map([
-          { challengeId: 'challenge_A_for_competence_1', result: 'ok' },
-          { challengeId: 'challenge_B_for_competence_1', result: 'ok' },
-          { challengeId: 'challenge_C_for_competence_1', result: 'ok' },
-          { challengeId: 'challenge_D_for_competence_2', result: 'ok' },
-          { challengeId: 'challenge_E_for_competence_2', result: 'ok' },
-          { challengeId: 'challenge_F_for_competence_2', result: 'ok' },
-          { challengeId: 'challenge_G_for_competence_3', result: 'ok' },
-          { challengeId: 'challenge_H_for_competence_3', result: 'ok' },
-          { challengeId: 'challenge_I_for_competence_3', result: 'ok' },
-          { challengeId: 'challenge_J_for_competence_4', result: 'ok' },
-          { challengeId: 'challenge_K_for_competence_4', result: 'ok' },
-          { challengeId: 'challenge_K_for_competence_4', result: 'ok' },
-          { challengeId: 'challenge_L_for_competence_4', result: 'ok' },
-        ], domainBuilder.buildAnswer);
-
-        const certificationAssessment = {
-          certificationAnswersByDate,
-          certificationChallenges: challenges,
-        };
-
-        sinon.stub(placementProfileService, 'getPlacementProfile').withArgs({
-          userId: certificationAssessment.userId,
-          limitDate: certificationAssessment.createdAt,
-          isV2Certification: certificationAssessment.isV2Certification,
-        }).resolves({ userCompetences });
-
-        // when
-        const error = await catchErr(certificationResultService.computeResult)({ certificationAssessment, continueOnError: false });
-
-        // then
-        expect(error).to.be.instanceOf(CertificationComputeError);
-        expect(error.message).to.equal('Plusieurs réponses pour une même épreuve');
-      });
-    });
-
-    context('When there are more challenges than answers ', () => {
-
-      it('should throw', async () => {
-        // given
-        const certificationAnswersByDate = _.map([
-          { challengeId: 'challenge_A_for_competence_1', result: 'ok' },
-          { challengeId: 'challenge_B_for_competence_1', result: 'ok' },
-          { challengeId: 'challenge_C_for_competence_1', result: 'ok' },
-          { challengeId: 'challenge_D_for_competence_2', result: 'ok' },
-          { challengeId: 'challenge_E_for_competence_2', result: 'ok' },
-          { challengeId: 'challenge_F_for_competence_2', result: 'ok' },
-          { challengeId: 'challenge_G_for_competence_3', result: 'ok' },
-          { challengeId: 'challenge_H_for_competence_3', result: 'ok' },
-          { challengeId: 'challenge_I_for_competence_3', result: 'ok' },
-          { challengeId: 'challenge_J_for_competence_4', result: 'ok' },
-          { challengeId: 'challenge_K_for_competence_4', result: 'ok' },
-        ], domainBuilder.buildAnswer);
-
-        const certificationAssessment = {
-          certificationAnswersByDate,
-          certificationChallenges: challenges,
-        };
-
-        sinon.stub(placementProfileService, 'getPlacementProfile').withArgs({
-          userId: certificationAssessment.userId,
-          limitDate: certificationAssessment.createdAt,
-          isV2Certification: certificationAssessment.isV2Certification,
-        }).resolves({ userCompetences });
-
-        // when
-        const error = await catchErr(certificationResultService.computeResult)({ certificationAssessment, continueOnError: false });
-
-        // then
-        expect(error).to.be.instanceOf(CertificationComputeError);
-        expect(error.message).to.equal('L’utilisateur n’a pas répondu à toutes les questions');
-      });
-    });
-
-    context('When there is no challenge for a competence ', () => {
-
-      it('should throw', async () => {
-        // given
-        const certificationAnswersByDate = _.map([
-          { challengeId: 'challenge_A_for_competence_1', result: 'ok' },
-          { challengeId: 'challenge_B_for_competence_1', result: 'ok' },
-          { challengeId: 'challenge_C_for_competence_1', result: 'ok' },
-          { challengeId: 'challenge_D_for_competence_2', result: 'ok' },
-          { challengeId: 'challenge_E_for_competence_2', result: 'ok' },
-          { challengeId: 'challenge_F_for_competence_2', result: 'ok' },
-          { challengeId: 'challenge_G_for_competence_3', result: 'ok' },
-          { challengeId: 'challenge_H_for_competence_3', result: 'ok' },
-          { challengeId: 'challenge_I_for_competence_3', result: 'ok' },
-          { challengeId: 'challenge_J_for_competence_4', result: 'ok' },
-          { challengeId: 'challenge_K_for_competence_4', result: 'ok' },
-        ], domainBuilder.buildAnswer);
-
-        const challenges = _.map([
-          { challengeId: 'challenge_A_for_competence_1', competenceId: 'competence_1', associatedSkillName: '@skillChallengeA_1', type: 'QCM' },
-          { challengeId: 'challenge_C_for_competence_1', competenceId: 'competence_1', associatedSkillName: '@skillChallengeC_1', type: 'QCM' },
-          { challengeId: 'challenge_B_for_competence_1', competenceId: 'competence_1', associatedSkillName: '@skillChallengeB_1', type: 'QCM' },
-          { challengeId: 'challenge_D_for_competence_2', competenceId: 'competence_2', associatedSkillName: '@skillChallengeD_2', type: 'QCM' },
-          { challengeId: 'challenge_E_for_competence_2', competenceId: 'competence_2', associatedSkillName: '@skillChallengeE_2', type: 'QCM' },
-          { challengeId: 'challenge_F_for_competence_2', competenceId: 'competence_2', associatedSkillName: '@skillChallengeF_2', type: 'QCM' },
-          { challengeId: 'challenge_G_for_competence_3', competenceId: 'competence_3', associatedSkillName: '@skillChallengeG_3', type: 'QCM' },
-          { challengeId: 'challenge_H_for_competence_3', competenceId: 'competence_3', associatedSkillName: '@skillChallengeH_3', type: 'QCM' },
-          { challengeId: 'challenge_I_for_competence_3', competenceId: 'competence_3', associatedSkillName: '@skillChallengeI_3', type: 'QCM' },
-        ], domainBuilder.buildCertificationChallengeWithType);
-
-        const certificationAssessment = {
-          certificationAnswersByDate,
-          certificationChallenges: challenges,
-        };
-
-        sinon.stub(placementProfileService, 'getPlacementProfile').withArgs({
-          userId: certificationAssessment.userId,
-          limitDate: certificationAssessment.createdAt,
-          isV2Certification: certificationAssessment.isV2Certification,
-        }).resolves({ userCompetences });
-
-        // when
-        const error = await catchErr(certificationResultService.computeResult)({ certificationAssessment, continueOnError: false });
-
-        // then
-        expect(error).to.be.instanceOf(CertificationComputeError);
-        expect(error.message).to.equal('Pas assez de challenges posés pour la compétence 4.4');
-      });
-    });
-
-    context('Compute certification result for jury (continue on error)', () => {
-      const continueOnError = true;
-
-      beforeEach(() => {
-        certificationAssessment = domainBuilder.buildCertificationAssessment({
-          ...certificationAssessmentData,
-          certificationAnswersByDate: wrongAnswersForAllChallenges(),
-          certificationChallenges: challenges,
-        });
-
-        sinon.stub(placementProfileService, 'getPlacementProfile').withArgs({
-          userId: certificationAssessment.userId,
-          limitDate: certificationAssessment.createdAt,
-          isV2Certification: certificationAssessment.isV2Certification,
-        }).resolves({ userCompetences });
-      });
-
-      it('should get user profile', async () => {
-        // when
-        await certificationResultService.computeResult({ certificationAssessment, continueOnError });
-
-        // then
-        sinon.assert.calledOnce(placementProfileService.getPlacementProfile);
-      });
-
-      context('when assessment is just started', () => {
-        let startedCertificationAssessment;
-
-        beforeEach(() => {
-          startedCertificationAssessment = domainBuilder.buildCertificationAssessment({
-            ...certificationAssessment,
-            completedAt: null,
-            state: states.STARTED,
-          });
-        });
-
-        it('should return list of competences with all certifiedLevel at -1', async () => {
-          // when
-          const certificationAssessmentScore = await certificationResultService.computeResult({
-            certificationAssessment: startedCertificationAssessment,
-            continueOnError,
-          });
-
-          // then
-          expect(certificationAssessmentScore.competenceMarks).to.deep.equal(expectedCertifiedCompetences);
-        });
-      });
-
-      context('when reproducibility rate is < 50%', () => {
-
-        it('should return list of competences with all certifiedLevel at -1', async () => {
-          // when
-          const certificationAssessmentScore = await certificationResultService.computeResult({ certificationAssessment, continueOnError });
-
-          // then
-          expect(certificationAssessmentScore.competenceMarks).to.deep.equal(expectedCertifiedCompetences);
-        });
-      });
-
-      context('when reproducibility rate is between 80% and 100%', () => {
-
-        beforeEach(() => {
-          certificationAssessment.certificationAnswersByDate = correctAnswersForAllChallenges();
-        });
-
-        it('should return list of competences with all certifiedLevel equal to estimatedLevel', async () => {
-          // given
-          const expectedCertifiedCompetences = [
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_1_1,
-              level: 1,
-              score: pixForCompetence1,
-            }),
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_2_2,
-              level: 2,
-              score: pixForCompetence2,
-            }),
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_3_3,
-              level: 3,
-              score: pixForCompetence3,
-            }),
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_4_4,
-              level: 4,
-              score: pixForCompetence4,
-            }),
-          ];
-
-          // when
-          const certificationAssessmentScore = await certificationResultService.computeResult({ certificationAssessment, continueOnError });
-
-          // then
-          expect(certificationAssessmentScore.competenceMarks).to.deep.equal(expectedCertifiedCompetences);
-        });
-
-        it('should return list of competences with certifiedLevel = estimatedLevel except for failed competence', async () => {
-          // given
-          certificationAssessment.certificationAnswersByDate = answersToHaveOnlyTheLastCompetenceFailed();
-          const expectedCertifiedCompetences = [
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_1_1,
-              level: 1,
-              score: pixForCompetence1,
-            }),
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_2_2,
-              level: 2,
-              score: pixForCompetence2,
-            }),
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_3_3,
-              level: 3,
-              score: pixForCompetence3,
-            }),
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_4_4,
-            })];
-
-          // when
-          const certificationAssessmentScore = await certificationResultService.computeResult({ certificationAssessment, continueOnError });
-
-          // then
-          expect(certificationAssessmentScore.competenceMarks).to.deep.equal(expectedCertifiedCompetences);
-        });
-      });
-
-      context('when reproducibility rate is between 50% and 80%', () => {
-
-        beforeEach(() => {
-          certificationAssessment.certificationAnswersByDate = answersWithReproducibilityRateLessThan80();
-        });
-
-        it('should return list of competences with certifiedLevel less or equal to estimatedLevel', async () => {
-          // given
-          const malusForFalseAnswer = 8;
-          const expectedCertifiedCompetences = [
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_1_1,
-              level: 0,
-              score: pixForCompetence1 - malusForFalseAnswer,
-            }),
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_2_2,
-              level: 2,
-              score: pixForCompetence2,
-            }),
-            competenceWithMarks_3_3,
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_4_4,
-              level: 3,
-              score: pixForCompetence4 - malusForFalseAnswer,
-            }),
-          ];
-
-          // when
-          const certificationAssessmentScore = await certificationResultService.computeResult({ certificationAssessment, continueOnError });
-
-          // then
-          expect(certificationAssessmentScore.competenceMarks).to.deep.equal(expectedCertifiedCompetences);
-        });
-
-        it('should return the percentage of correct answers', async () => {
-          // given
-          const certificationAssessmentWithNeutralizedChallenge = _.cloneDeep(certificationAssessment);
-          certificationAssessmentWithNeutralizedChallenge.certificationChallenges[0].isNeutralized = true;
-
-          // when
-          const { percentageCorrectAnswers } = await certificationResultService.computeResult({ certificationAssessment: certificationAssessmentWithNeutralizedChallenge, continueOnError });
-
-          // then
-          expect(percentageCorrectAnswers).to.deep.equal(64);
-        });
-
-        context('when one competence is evaluated with 3 challenges', () => {
-
-          context('with one OK, one KO and one QROCM-dep OK', () => {
-
-            it('should return level obtained equal to level positioned minus one', async () => {
-              // Given
-              const positionedLevel = 2;
-              const positionedScore = 20;
-
-              const answers = _.map([
-                { challengeId: 'challenge_A_for_competence_1', result: 'ok' },
-                { challengeId: 'challenge_B_for_competence_1', result: 'ok' },
-                { challengeId: 'challenge_C_for_competence_1', result: 'ko' },
-              ], domainBuilder.buildAnswer);
-
-              const challenges = _.map([
-                { challengeId: 'challenge_A_for_competence_1', competenceId: 'competence_1', associatedSkillName: '@skillChallengeA_1' },
-                { challengeId: 'challenge_B_for_competence_1', competenceId: 'competence_1', associatedSkillName: '@skillChallengeB_1' },
-                { challengeId: 'challenge_C_for_competence_1', competenceId: 'competence_1', associatedSkillName: '@skillChallengeC_1' },
-              ], domainBuilder.buildCertificationChallengeWithType);
-
-              const userCompetences = [
-                _buildUserCompetence(competence_1, positionedScore, positionedLevel),
-              ];
-
-              certificationAssessment.certificationAnswersByDate = answers;
-              certificationAssessment.certificationChallenges = challenges;
-              placementProfileService.getPlacementProfile.restore();
-              sinon.stub(placementProfileService, 'getPlacementProfile').withArgs({
-                userId: certificationAssessment.userId,
-                limitDate: certificationAssessment.createdAt,
-                isV2Certification: certificationAssessment.isV2Certification,
-              }).resolves({ userCompetences });
-
-              // When
-              const certificationAssessmentScore = await certificationResultService.computeResult({ certificationAssessment, continueOnError });
-
-              // Then
-              expect(certificationAssessmentScore.competenceMarks[0].level).to.deep.equal(positionedLevel - 1);
-              expect(certificationAssessmentScore.competenceMarks[0].score).to.deep.equal(positionedScore - 8);
-            });
-          });
-        });
-      });
-    });
-
-    context('Calculate certification result when assessment is completed (stop on error)', () => {
-      const continueOnError = false;
-
-      beforeEach(() => {
-        certificationAssessment = domainBuilder.buildCertificationAssessment({
-          ...certificationAssessmentData,
-          certificationAnswersByDate: wrongAnswersForAllChallenges(),
-          certificationChallenges: challenges,
-        });
-        sinon.stub(placementProfileService, 'getPlacementProfile').withArgs({
-          userId: certificationAssessment.userId,
-          limitDate: certificationAssessment.createdAt,
-          isV2Certification: certificationAssessment.isV2Certification,
-        }).resolves({ userCompetences });
-      });
-
-      context('when reproducibility rate is < 50%', () => {
-
-        it('should return list of competences with all certifiedLevel at -1', async () => {
-          // when
-          const certificationAssessmentScore = await certificationResultService.computeResult({ certificationAssessment, continueOnError });
-
-          // then
-          expect(certificationAssessmentScore.competenceMarks).to.deep.equal(expectedCertifiedCompetences);
-        });
-      });
-
-      context('when reproducibility rate is between 80% and 100%', () => {
-
-        beforeEach(() => {
-          certificationAssessment.certificationAnswersByDate = correctAnswersForAllChallenges();
-        });
-
-        it('should return list of competences with all certifiedLevel equal to estimatedLevel', async () => {
-          // given
-          const expectedCertifiedCompetences = [
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_1_1,
-              level: 1,
-              score: pixForCompetence1,
-            }),
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_2_2,
-              level: 2,
-              score: pixForCompetence2,
-            }),
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_3_3,
-              level: 3,
-              score: pixForCompetence3,
-            }),
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_4_4,
-              level: 4,
-              score: pixForCompetence4,
-            }),
-          ];
-
-          // when
-          const certificationAssessmentScore = await certificationResultService.computeResult({ certificationAssessment, continueOnError });
-
-          // then
-          expect(certificationAssessmentScore.competenceMarks).to.deep.equal(expectedCertifiedCompetences);
-        });
-
-        it('should return list of competences with certifiedLevel = estimatedLevel except for failed competence', async () => {
-          // given
-          certificationAssessment.certificationAnswersByDate = answersToHaveOnlyTheLastCompetenceFailed();
-          const expectedCertifiedCompetences = [
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_1_1,
-              level: 1,
-              score: pixForCompetence1,
-            }),
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_2_2,
-              level: 2,
-              score: pixForCompetence2,
-            }),
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_3_3,
-              level: 3,
-              score: pixForCompetence3,
-            }),
-            competenceWithMarks_4_4,
-          ];
-
-          // when
-          const certificationAssessmentScore = await certificationResultService.computeResult({ certificationAssessment, continueOnError });
-
-          // then
-          expect(certificationAssessmentScore.competenceMarks).to.deep.equal(expectedCertifiedCompetences);
-        });
-      });
-
-      context('when reproducibility rate is between 50% and 80%', () => {
-
-        beforeEach(() => {
-          certificationAssessment.certificationAnswersByDate = answersWithReproducibilityRateLessThan80();
-        });
-
-        it('should return list of competences with certifiedLevel less or equal to estimatedLevel', async () => {
-          // given
-          const malusForFalseAnswer = 8;
-          const expectedCertifiedCompetences = [
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_1_1,
-              level: 0,
-              score: pixForCompetence1 - malusForFalseAnswer,
-            }),
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_2_2,
-              level: 2,
-              score: pixForCompetence2,
-            }),
-            competenceWithMarks_3_3,
-            domainBuilder.buildCompetenceMark({
-              ...competenceWithMarks_4_4,
-              level: 3,
-              score: pixForCompetence4 - malusForFalseAnswer,
-            }),
-          ];
-
-          // when
-          const certificationAssessmentScore = await certificationResultService.computeResult({ certificationAssessment, continueOnError });
-
-          // then
-          expect(certificationAssessmentScore.competenceMarks).to.deep.equal(expectedCertifiedCompetences);
-        });
-      });
-
-      context('when only one challenge is asked for a competence', () => {
-
-        it('certifies a level below the estimated one if reproducibility rate is < 70%', async () => {
-          // given
-          const userCompetences = [
-            _buildUserCompetence(competence_5, 50, 5),
-            _buildUserCompetence(competence_6, 36, 3),
-          ];
-          certificationAssessment.certificationChallenges = _.map([
-            {
-              challengeId: 'challenge_A_for_competence_5',
-              competenceId: 'competence_5',
-              associatedSkillName: '@skillChallengeA_5',
-            },
-            {
-              challengeId: 'challenge_A_for_competence_6',
-              competenceId: 'competence_6',
-              associatedSkillName: '@skillChallengeA_6',
-            },
-            {
-              challengeId: 'challenge_B_for_competence_6',
-              competenceId: 'competence_6',
-              associatedSkillName: '@skillChallengeB_6',
-            },
-            {
-              challengeId: 'challenge_C_for_competence_6',
-              competenceId: 'competence_6',
-              associatedSkillName: '@skillChallengeC_6',
-            },
-          ], domainBuilder.buildCertificationChallengeWithType);
-
-          placementProfileService.getPlacementProfile.restore();
-          sinon.stub(placementProfileService, 'getPlacementProfile').withArgs({
-            userId: certificationAssessment.userId,
-            limitDate: certificationAssessment.createdAt,
-            isV2Certification: certificationAssessment.isV2Certification,
-          }).resolves({ userCompetences });
-
-          certificationAssessment.certificationAnswersByDate = _.map([
-            ({ challengeId: 'challenge_A_for_competence_5', result: 'ok' }),
-            ({ challengeId: 'challenge_A_for_competence_6', result: 'ko' }),
-            ({ challengeId: 'challenge_B_for_competence_6', result: 'ok' }),
-            ({ challengeId: 'challenge_C_for_competence_6', result: 'ko' }),
-          ], domainBuilder.buildAnswer);
-
-          const expectedCertifiedCompetences = [
-            domainBuilder.buildCompetenceMark({
-              competence_code: '5.5',
-              area_code: '5',
-              competenceId: 'competence_5',
-              level: 4,
-              score: 40,
-            }),
-            domainBuilder.buildCompetenceMark({
-              competence_code: '6.6',
-              area_code: '6',
-              competenceId: 'competence_6',
-              level: UNCERTIFIED_LEVEL,
-              score: 0,
-            }),
-          ];
-
-          // when
-          const certificationAssessmentScore = await certificationResultService.computeResult({ certificationAssessment, continueOnError });
-
-          // then
-          expect(certificationAssessmentScore.competenceMarks).to.deep.equal(expectedCertifiedCompetences);
-        });
-      });
-
-      context('when challenges contains one QROCM-dep challenge to validate two skills', () => {
-
-        beforeEach(() => {
-          const userCompetences = [
-            _buildUserCompetence(competence_5, 50, 5),
-            _buildUserCompetence(competence_6, 36, 3),
-          ];
-
-          certificationAssessment.certificationChallenges = _.map([
-            {
-              challengeId: 'challenge_A_for_competence_5',
-              competenceId: 'competence_5',
-              associatedSkillName: '@skillChallengeA_5',
-              type: 'QCM',
-            },
-            {
-              challengeId: 'challenge_B_for_competence_5',
-              competenceId: 'competence_5',
-              associatedSkillName: '@skillChallengeB_5',
-              type: 'QROCM-dep',
-            },
-            {
-              challengeId: 'challenge_A_for_competence_6',
-              competenceId: 'competence_6',
-              associatedSkillName: '@skillChallengeA_6',
-              type: 'QCM',
-            },
-            {
-              challengeId: 'challenge_B_for_competence_6',
-              competenceId: 'competence_6',
-              associatedSkillName: '@skillChallengeB_6',
-              type: 'QCM',
-            },
-            {
-              challengeId: 'challenge_C_for_competence_6',
-              competenceId: 'competence_6',
-              associatedSkillName: '@skillChallengeC_6',
-              type: 'QCM',
-            },
-          ], domainBuilder.buildCertificationChallengeWithType);
-
-          placementProfileService.getPlacementProfile.restore();
-          sinon.stub(placementProfileService, 'getPlacementProfile').withArgs({
-            userId: certificationAssessment.userId,
-            limitDate: certificationAssessment.createdAt,
-            isV2Certification: certificationAssessment.isV2Certification,
-          }).resolves({ userCompetences });
-
-        });
-
-        it('should compute the result as if QROCM-dep was two OK challenges', async () => {
-          // given
-          certificationAssessment.certificationAnswersByDate = _.map([
-            ({ challengeId: 'challenge_A_for_competence_5', result: 'ok' }),
-            ({ challengeId: 'challenge_B_for_competence_5', result: 'ok' }),
-            ({ challengeId: 'challenge_A_for_competence_6', result: 'ko' }),
-            ({ challengeId: 'challenge_B_for_competence_6', result: 'ok' }),
-            ({ challengeId: 'challenge_C_for_competence_6', result: 'ko' }),
-          ], domainBuilder.buildAnswer);
-
-          const expectedCertifiedCompetences = [
-            domainBuilder.buildCompetenceMark({
-              competence_code: '5.5',
-              area_code: '5',
-              competenceId: 'competence_5',
-              level: 5,
-              score: 40,
-            }),
-            domainBuilder.buildCompetenceMark({
-              competence_code: '6.6',
-              area_code: '6',
-              competenceId: 'competence_6',
-              level: UNCERTIFIED_LEVEL,
-              score: 0,
-            }),
-          ];
-
-          // when
-          const certificationAssessmentScore = await certificationResultService.computeResult({ certificationAssessment, continueOnError });
-
-          // then
-          expect(certificationAssessmentScore.competenceMarks).to.deep.equal(expectedCertifiedCompetences);
-        });
-
-        it('should compute the result of QROCM-dep as only one OK because result is partially right', async () => {
-          // given
-          certificationAssessment.certificationAnswersByDate = _.map([
-            { challengeId: 'challenge_A_for_competence_5', result: 'ok' },
-            { challengeId: 'challenge_B_for_competence_5', result: 'partially' },
-            { challengeId: 'challenge_A_for_competence_6', result: 'ko' },
-            { challengeId: 'challenge_B_for_competence_6', result: 'ok' },
-            { challengeId: 'challenge_C_for_competence_6', result: 'ok' },
-          ], domainBuilder.buildAnswer);
-
-          const expectedCertifiedCompetences = [
-            domainBuilder.buildCompetenceMark({
-              competence_code: '5.5',
-              area_code: '5',
-              competenceId: 'competence_5',
-              level: 4,
-              score: 40,
-            }),
-            domainBuilder.buildCompetenceMark({
-              competence_code: '6.6',
-              area_code: '6',
-              competenceId: 'competence_6',
-              level: 2,
-              score: 28,
-            }),
-          ];
-
-          // when
-          const certificationAssessmentScore = await certificationResultService.computeResult({ certificationAssessment, continueOnError });
-
-          // then
-          expect(certificationAssessmentScore.competenceMarks).to.deep.equal(expectedCertifiedCompetences);
-        });
-      });
-
-      context('when there are challenges for non-certifiable competences', () => {
-        let challenges;
-
-        beforeEach(() => {
-          challenges = _.map([
-            { challengeId: 'challenge_A_for_competence_1', competenceId: 'competence_1', associatedSkillName: '@skillChallengeA_1' },
-            { challengeId: 'challenge_M_for_competence_5', competenceId: 'competence_5', associatedSkillName: '@skillChallengeM_5' },
-            { challengeId: 'challenge_N_for_competence_6', competenceId: 'competence_6', associatedSkillName: '@skillChallengeN_6' },
-          ], domainBuilder.buildCertificationChallengeWithType);
-          certificationAssessment.certificationChallenges = challenges;
-
-          certificationAssessment.certificationAnswersByDate = _.map([
-            { challengeId: 'challenge_A_for_competence_1', result: 'ko' },
-
-            { challengeId: 'challenge_M_for_competence_5', result: 'ok' },
-            { challengeId: 'challenge_N_for_competence_6', result: 'ok' },
-          ], domainBuilder.buildAnswer);
-        });
-
-        it('should not include the extra challenges when computing reproducibility', async () => {
-          // when
-          const certificationAssessmentScore = await certificationResultService.computeResult({ certificationAssessment, continueOnError });
-
-          // then
-          expect(certificationAssessmentScore.percentageCorrectAnswers).to.equal(0);
-        });
-      });
-    });
-  });
-});
